The configuration file we want to see. Allow silent local message pass-through to specific servers.
The use-case here is when one mini-game server implements team chat itself, with team
prefixes and perhaps spectator chat, with recipients not seeing chat from opposing teams
or spectators, as appropriate. BungeeChat2 can't know about this arrangement, so it's
sensible to hand it over to the local server, without formatting and sending messages to any
of the connected clients.

Global, staff and helpop channels remain available, as does private messaging, but these
being commands can be controlled using context-based permissions.