Use a normal `WeakHashMap`
`new MapMaker().weakValues().makeMap()` returns a concurrent weak map here. Which is abolutely not desired as the synchronization methods may interfere. That's why I chose to use a normal `WeakHashMap` instead.
Also did some formatting and clean up. For example you checked 3 times if the `connection` is (not) `null`. Checking once and storing in a variable is much better.
And the formatting was mostly about leaving space, and using consistent formatting of the conditions.
Lastly some finals to indicate that we will not change the variables.