Back to the MapMaker variant as that is cleaner to read
There may be issues here as `ConcurrentMap`s and `synchronnized` supposedly don't mix well. However I believe this to be fine here. We don't use the concurrency of the map ever so it shouldn't matter.
Another concern is that the map may be performance heavy.

If any of these concerns turn out to be substancial, we can always just revert this commit.